refactor Game.jsx - put music in dedicated component
