+++ src/client/pages/game/Game.jsx
@@ -5,43 +5,19 @@ import { useTetrisGame } from "hooks";
 import { Link } from "react-router-dom";
 import { GameContext } from "store";
 import NextPieces from "./NextPieces";
-import TetrisTheme from "assets/music/Tetris_theme.ogg";
-import TetrisGameOverTheme from "assets/music/Tetris_game_over.ogg";
 
 import useEventListener from "hooks/useEventListener";
 import useThrottle from "hooks/useThrottle";
 import { DEFAULT_REPEAT_TIMEOUT } from "hooks/tetris-game/constants";
+import AudioTheme from "./AudioTheme";
 
 export default function Game() {
   const { state, dispatch } = React.useContext(GameContext);
   const { movePiece } = useTetrisGame(10, 20);
+
+  // Add keyboard event
   const throttledMove = useThrottle(movePiece, DEFAULT_REPEAT_TIMEOUT);
   useEventListener("keydown", throttledMove);
-
-  const audioRef = React.useRef(null);
-
-  React.useEffect(() => {
-    if (state.alive) {
-      audioRef.current.src = TetrisTheme;
-      audioRef.current.volume = 0.2;
-      audioRef.current.loop = true;
-      audioRef.current.play();
-    } else {
-      audioRef.current.loop = false;
-      audioRef.current.src = TetrisGameOverTheme;
-      audioRef.current.playbackRate = 1.0;
-      audioRef.current.volume = 1.0;
-      audioRef.current.play();
-    }
-
-    return () => {
-      audioRef.current.src = null;
-    };
-  }, [state.alive]);
-
-  React.useEffect(() => {
-    audioRef.current.playbackRate = state.speedRate;
-  }, [state.speedRate]);
 
   return (
     <FlexBox
@@ -50,7 +26,7 @@ export default function Game() {
       height={"full"}
       className="justify-center items-center"
     >
-      <audio ref={audioRef} />
+      <AudioTheme alive={state.alive} speedRate={state.speedRate} />
       <Link
         to="/"
         className="font-semibold border-2 p-2 mb-2 border-red-300 rounded"
